--- frontend/app/(home)/components/UseCases/components/UseCasesListing/data/cases.ts
@@ -1,49 +1,0 @@
-import { UseCase } from "../types";
-
-export const casesExamples: UseCase[] = [
-  {
-    id: "1",
-    name: "Research and Studies",
-    description: "Quivr is your indispensable research companion.",
-    discussions: [
-      {
-        question: "Can you explain what can Research and Studies do ?",
-        answer: `Hey there ! First you can upload documents by clicking on “+”. Then you can ask questions to your documents, and interact with them. `,
-      },
-      {
-        question: "Cool, and what about brains ? ",
-        answer: `Brains are like your data base. You can feed a brain with multiple documents and then interact with one brain. You can also talk to Chat GPT by typing @ChatGPT4 `,
-      },
-    ],
-  },
-  {
-    id: "2",
-    name: "Legal research",
-    description: "Your ultimate digital ally in the field of law.",
-    discussions: [
-      {
-        question: "Can you explain whar can Legal research do ?",
-        answer: `Hey there ! First you can upload documents by clicking on “+”. Then you can ask questions to your documents, and interact with them. `,
-      },
-      {
-        question: "Cool, and what about brains ? ",
-        answer: `Brains are like your data base. You can feed a brain with multiple documents and then interact with one brain. You can also talk to Chat GPT by typing @ChatGPT4 `,
-      },
-    ],
-  },
-  {
-    id: "3",
-    name: "Sales",
-    description: "Placeholder",
-    discussions: [
-      {
-        question: "Can you explain what can Sales do ?",
-        answer: `Hey there ! First you can upload documents by clicking on “+”. Then you can ask questions to your documents, and interact with them. `,
-      },
-      {
-        question: "Cool, and what about brains ? ",
-        answer: `Brains are like your data base. You can feed a brain with multiple documents and then interact with one brain. You can also talk to Chat GPT by typing @ChatGPT4 `,
-      },
-    ],
-  },
-];
