--- frontend/app/(home)/components/UseCases/components/UseCasesListing/types.ts
@@ -1,9 +1,0 @@
-export type UseCase = {
-  id: string;
-  name: string;
-  description: string;
-  discussions: {
-    question: string;
-    answer: string;
-  }[];
-};
