+++ frontend/lib/types/UseCase.ts
@@ -1,0 +1,11 @@
+type Discussion = {
+  question: string;
+  answer: string;
+};
+
+export type UseCase = {
+  id: string;
+  name: string;
+  description: string;
+  discussions: Discussion[];
+};
